model-training/requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Training and Optimization
trl>=0.7.0
datasets>=2.14.0
scipy>=1.11.0
numpy>=1.24.0

# RL Libraries
gym>=0.26.0
stable-baselines3>=2.2.0

# Data and Storage
cloudpathlib[s3]>=0.16.0
aioboto3>=12.0.0
aiohttp>=3.9.0

# Evaluation and Metrics
scikit-learn>=1.3.0
pandas>=2.0.0

# Logging and Monitoring
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
omegaconf>=2.3.0

# HuggingFace Hub
huggingface-hub>=0.19.0

# Utilities
psutil>=5.9.0
colorlog>=6.8.0
bittensor>=7.0.0
orjson>=3.9.0
tabulate>=0.9.0
aiofiles>=23.0.0
